refactor: 불필요한 녹음 config 삭제 및 영상 녹화 config 파일명 수정
- utils/recorder.ts -> utils/videoRecorderConfig.ts 으로 영상 녹화 config임을 직관적으로 파악 가능하도록 수정
